fix c grammar for wide string literals
It seems to me that it's impossible to parse wide string literals using c.peg.  (i.e. L"foo").  The issue appears to be that the "L" will be captured by the Identifier subrule in PrimaryExpression before StringLiteral can capture it.  My fix is to reorder these subrules, putting StringLiteral before Identifier.