Bugfix: call node print function instead of built-in print function (#188)
When printing the AST to a string using the SprintSyntaxTree function, I was seeing pointer addresses being printed to stdout. Upon looking at the code, I found that in the `printFunc` function that writes the AST to a `io.Writer`, the recursive call is incorrectly calling the built-in `print` function instead of `printFunc`.